See description
1. Moving sounds to centralized location (root/assets)
2. Deleting all png and ogg files from android's assets folders, since
these are generated by the Gradle build script
3. Implementing new SoundEngine for iOS, which allows me to play wav
files instead of caf/aifc and not rely on the Ringer/Sound Effects
volume control for my sound effects separately from the background
music.
4. Implementing separate MUSIC_RESUME def to explicitly resume whatever
music was playing before. Calling MUSIC_PLAY_* will always instantiate
a new track and play it from the beginning

+++ src/core/game/ui/Renderer.cpp
@@ -672,7 +672,7 @@ void Renderer::renderHud(Game& game, BackButton &backButton)
     std::stringstream ss;
     ss << game.getNumTotalCarrots();
     std::string totalNumGoldenCarrots = ss.str();
-    int offset = 1 + totalNumGoldenCarrots.size();
+    int offset = 1 + (int) totalNumGoldenCarrots.size();
     
     /// Render Num / Total Carrots
